+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.1.333 (2024-4-16)
 
-- chore: embedding discuss data
+- chore: embedding discuss data by type board and id
 
 ## 0.1.332 (2024-4-15)